requirements.txt
Flask
requests
pdfplumber
python-docx
python-dotenv